project two files :)
also a package with two players i'm working on :) currently both are /technically/ the same because its one file i duplicated because i thought the result (collegestudent.java) was funny and wanted to preserve it. the collegestudent player wins but only because it drives itself into debt. i made a mistake somewhere